WIP: Organize public APIs in lib.rs
