+++ .bazelrc
@@ -9,3 +9,5 @@ test --test_verbose_timeout_warnings
 
 build --aspects=@rules_rust//rust:defs.bzl%rust_clippy_aspect
 build --output_groups=+clippy_checks
+
+common --noenable_bzlmod
